+++ src/lib/prisma/prisma.ts
@@ -1,0 +1,3 @@
+import {PrismaClient} from "@prisma/client";
+
+export const prisma = new PrismaClient()
